secure_ai_layer/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
PyYAML>=6.0.2
pytesseract>=0.3.10
Pillow>=10.2.0
PyMuPDF>=1.24.0
spacy>=3.7.0
Jinja2>=3.1.3
watchdog>=4.0.0
openai>=1.14.0
sentence-transformers>=2.7.0
redis>=5.0.3
WeasyPrint>=62.0
pytest>=8.0.0
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.6.0
